new notes are random color
